Removed warnings

+++ src/core/Components/RotationComponent.hpp
@@ -78,10 +78,10 @@ namespace Core
 		static RotationComponent GetComponentRotateAxis(float angleRad, float axisX, float axisY, float axisZ)
 		{
 			RotationComponent rc;
-			rc.rotation[0] = axisX * std::sin(angleRad * 0.5);
-			rc.rotation[1] = axisY * std::sin(angleRad * 0.5);
-			rc.rotation[2] = axisZ * std::sin(angleRad * 0.5);
-			rc.rotation[3] = std::cos(angleRad * 0.5);
+			rc.rotation[0] = axisX * (float)std::sin((double)angleRad * 0.5);
+			rc.rotation[1] = axisY * (float)std::sin((double)angleRad * 0.5);
+			rc.rotation[2] = axisZ * (float)std::sin((double)angleRad * 0.5);
+			rc.rotation[3] = (float)std::cos((double)angleRad * 0.5);
 
 			return rc;
 		}
